// src/modules/safety/permissions/interfaces/permissionData.interface.ts
export interface permissionData {
  idRol?: string;
  idModule?: string;
  permission?: number;
  pinPermission?: string;
  permissionData?: number[];
  rolName: string;
  rolLevel: Number;
  moduleName: string;
  menuComponent: string;
  menuVideo: string;
  menuFather: string;
  menuOrder: number;
  menuIcon: string;
}